(PC-15975) refactor(analytics): simplfy the analytics file
