+++ src/features/navigation/helpers/openUrl.ts
@@ -4,7 +4,7 @@ import { Alert, Linking } from 'react-native'
 import { getScreenFromDeeplink } from 'features/deeplinks/getScreenFromDeeplink'
 import { navigateFromRef } from 'features/navigation/navigationRef'
 import { analytics } from 'libs/firebase/analytics'
-import { OfferAnalyticsData } from 'libs/firebase/analytics/analytics'
+import { OfferAnalyticsData } from 'libs/firebase/analytics/types'
 import { captureMonitoringError } from 'libs/monitoring'
 
 import { isAppUrl } from './isAppUrl'
